Update correct structure.

+++ SpellVisualMapBuilder/CreatureModelData.cs
@@ -149,7 +149,7 @@ namespace SpellEditor.Sources.DBC
             // This is always 0. It would be used, if something was in here. Its pushed into:
             //   M2Scene::AddNewModel(GetM2Cache(), Modelpath, AlternateModel, 0)
             public UInt32 AlternateModel;
-            public UInt32 sizeClass;
+            public float sizeClass;
             public UInt32 modelScale; // Documented as float, IS NOT
             public UInt32 BloodLevel;
             public UInt32 Footprint;
